Split out total uplift test
